Partial commit, work on memory allocation in Surface
git-svn-id: svn://svn.code.sf.net/p/vmsvga2/code/VMsvga2/trunk@146 f31ded45-fcaa-4108-8aac-ff0a0b8cda5a

+++ FB/SVGADevice.cpp
@@ -786,7 +786,7 @@ bool CLASS::defineGMR2(uint32_t gmrId, uint32_t numPages)
 }
 
 bool CLASS::remapGMR2(uint32_t gmrId, uint32_t flags, uint32_t offsetPages,
-					  uint32_t numPages, void const* suffix, uint32_t suffixSize)
+					  uint32_t numPages, void const* suffix, size_t suffixSize)
 {
 	if (!suffix && suffixSize)
 		return false;
